requirements.txt
waveshare-epaper==1.3.0
pillow==10.4.0
gpiozero
fastapi~=0.111.0
fastapi-cli==0.0.4
uvicorn==0.30.1
ngrok~=1.3.0
python-dotenv~=1.0.1
pytest~=8.2.2
pydantic~=2.8.2
twilio~=9.2.3